requirements.txt
opencv-python-headless
numpy
pandas
moviepy
SpeechRecognition
yt_dlp
scikit-learn
matplotlib
seaborn
openai-whisper
streamlit
plotly
joblib